requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.6
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.4
django-phonenumber-field==8.1.0
django-widget-tweaks==1.5.0
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.2
fonttools==4.59.0
html5lib==1.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
oscrypto==1.3.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
phonenumberslite==9.0.10
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rpds-py==0.27.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
yarg==0.1.9
zopfli==0.2.3.post1
